cand-master/requirements.txt
fastapi
mysql-connector-python
python-dotenv
pydantic
